Move config out to its own module
